Sped up/reduced memory usage in GetElasticPartials.m
I now moved the UNprojection of calculated partials inside the for
loop as opposed to doing it for all segments at once. This finally
results in no paging out for the global test model, with a speedup
of about 2x (the previous version took 211 seconds on my machine and
this new version takes 100 seconds).