Sped up/reduced memory usage for GetElasticPartials.m
I moved the unprojection of calculated partials inside the for loop
as opposed to the previous version in which the unprojection was done
for all segments at once. On my machine, this allows partials to be
calculated for the global model without using all 32 GB of RAM, and
as a result, runtime is halved (from 200 seconds to 100 seconds).